// src/lib/api/schemas/variantPrice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { PriceType } from './priceType.ts';
import type { VariantId } from './variantId.ts';
import type { VariantPriceId } from './variantPriceId.ts';
import type { VariantStatus } from './variantStatus.ts';

export interface VariantPrice {
  amount: number;
  created_at: string;
  currency: string;
  id: VariantPriceId;
  price_type: PriceType;
  status: VariantStatus;
  updated_at: string;
  variant_id: VariantId;
}